Creation of first objects
Creation of grid ans snake object in js. I created the methode to move the snake for one step, verfiying if it's a game over or not (the snake crash himself). I made a game over if the snake go out of the grid, but it's a mistake, a gird is a torus in this game.